requirements.txt
Pillow
PyScreeze
